requirements.txt
asgiref==3.11.1
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt
drf-spectacular
sqlparse==0.5.5
celery
redis
psycopg2